Updated to correct Unity version, changed Mopuse Movement to float for greater control of speed

+++ Assets/Scripts/PlayerController.cs
@@ -7,7 +7,7 @@ using UnityEngine.UI;
 public class PlayerController : MonoBehaviour
 {
     //Startup Variables
-    public int mouseSpeed;
+    public float mouseSpeed;
     public int horizontalRotation;
     public int verticalRotation;
     public int VerticalOffset;
